+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Dates should be in`YYYY-MM-DD` format and versions are in [semantic versioning](
 
 - Added bidding scholar balance to submission bidding table.
 - Fixed #51 sticky header.
+- Saving feedback in sticky header.
 
 ### Maintenance